Extract CLI args tests
